Move bridge client sessions into a dedicated owner
Give one instance manager ownership of connected streams, handler tasks, request framing coordination, and session cleanup. Keep the server responsible for listener acceptance, server cancellation, shutdown orchestration, and lifecycle events while preserving every handshake and task transition.
