+++ Packages/src/Editor/Infrastructure/UnityCliLoopBridgeClientSessionManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e62d177cf2934c958c6f86185877f86b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
